Create function override for testing, use hostfs and clean up ioctl error handling
